+++ esp_cam_sensor/sensors/pivariety/include/pivariety_types.h
@@ -1,0 +1,25 @@
+/*
+ * SPDX-FileCopyrightText: 2024-2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*
+ * pivariety camera sensor register type definition.
+ */
+typedef struct {
+    uint16_t reg;
+    uint32_t val;
+} pivariety_reginfo_t;
+
+#ifdef __cplusplus
+}
+#endif